requirements.txt
numpy==1.24.4
matplotlib==3.7.5
py-cpuinfo==9.0.0